requirements.txt
aiosqlite==0.21.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.11.12
chardet==3.0.4
charset-normalizer==3.4.4
deep-translator==1.11.4
eng_to_ipa==0.0.2
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
iso8601==2.1.0
pypika-tortoise==0.6.2
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.5
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.8
tortoise-orm==0.25.1
typing_extensions==4.15.0
urllib3==2.5.0
